move sparque functionality into sparque feature extension
* this way the feature state works with feature flag
