+++ src/app/extensions/sparque/models/sparque-config/sparque-config.model.ts
