+++ src/app/extensions/sparque/models/sparque/sparque.interface.ts
