+++ src/app/extensions/sparque/services/sparque-api/sparque-api.service.ts
@@ -1,17 +1,17 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { Store, select, createSelector } from '@ngrx/store';
+import { Store, createSelector, select } from '@ngrx/store';
 import { Observable, auditTime, combineLatest, map, take, tap } from 'rxjs';
 
 import { getCurrentLocale } from 'ish-core/store/core/configuration';
 import { getCurrentBasket } from 'ish-core/store/customer/basket';
 import { getLoggedInUser } from 'ish-core/store/customer/user';
+import { log } from 'ish-core/utils/dev/operators';
 import { URLFormParams } from 'ish-core/utils/url-form-params';
 
-import { environment } from '../../../../environments/environment';
-import { getSparqueState } from 'ish-core/store/shopping/sparque/sparque-store';
-import { getSparqueConfig, getSparqueConfigEndpoint } from 'ish-core/store/shopping/sparque/sparque-config';
-import { log } from 'ish-core/utils/dev/operators';
+import { environment } from '../../../../../environments/environment';
+import { getSparqueConfigEndpoint } from '../../store/sparque-config';
+import { getSparqueState } from '../../store/sparque-store';
 
 export const DEFINED_FACETS = ['category', 'brand'];
 
@@ -47,8 +47,8 @@ export class SparqueApiService {
     //const getSparqueConfig = createSelector(getSparqueState, state => state?.sparqueConfig);
     const getSparqueConfig2 = createSelector(getSparqueState, state => state?.sparqueConfig);
     console.log(this.store.pipe(select(getSparqueConfigEndpoint)));
-    this.store.pipe(select(getSparqueConfigEndpoint)).subscribe(data => console.log(data))
-    this.store.pipe(select(getSparqueConfigEndpoint),log("testdori")).subscribe()
+    this.store.pipe(select(getSparqueConfigEndpoint)).subscribe(data => console.log(data));
+    this.store.pipe(select(getSparqueConfigEndpoint), log('testdori')).subscribe();
     return this.httpClient.get<T>(`${environment.sparqueBaseURL}/${path.replace('//', '/')}`);
   }
