+++ src/app/extensions/sparque/sparque.module.ts
@@ -1,0 +1,10 @@
+import { NgModule } from '@angular/core';
+
+import { SharedModule } from 'ish-shared/shared.module';
+
+@NgModule({
+  imports: [SharedModule],
+  declarations: [],
+  exports: [SharedModule],
+})
+export class SparqueModule {}
