+++ src/app/extensions/sparque/store/sparque-config/index.ts
