+++ src/app/extensions/sparque/store/sparque-config/sparque-config.actions.ts
@@ -2,7 +2,7 @@ import { createAction } from '@ngrx/store';
 
 import { payload } from 'ish-core/utils/ngrx-creators';
 
-import { SparqueConfig } from 'ish-core/models/sparque/sparque-config.model';
+import { SparqueConfig } from '../../models/sparque-config/sparque-config.model';
 
 export const loadSparqueConfig = createAction('[Sparque Config Internal] Load Sparque Config');
