+++ src/app/extensions/sparque/store/sparque-store.ts
@@ -1,8 +1,8 @@
 import { createFeatureSelector } from '@ngrx/store';
-import { SparqueConfig } from 'ish-core/models/sparque/sparque-config.model';
+
+import { SparqueConfig } from '../models/sparque-config/sparque-config.model';
 
 export interface SparqueState {
   sparqueConfig: SparqueConfig;
 }
-
 export const getSparqueState = createFeatureSelector<SparqueState>('sparque');
