+++ src/environments/environment.model.ts
@@ -5,8 +5,8 @@ import { DataRetentionPolicy } from 'ish-core/utils/meta-reducers';
 import { MultiSiteLocaleMap } from 'ish-core/utils/multi-site/multi-site.service';
 
 import { AddressDoctorConfig } from '../app/extensions/address-doctor/models/address-doctor/address-doctor-config.model';
+import { SparqueConfig } from '../app/extensions/sparque/models/sparque-config/sparque-config.model';
 import { TactonConfig } from '../app/extensions/tacton/models/tacton-config/tacton-config.model';
-import { SparqueConfig } from 'ish-core/models/sparque/sparque-config.model';
 
 export interface Environment {
   /* INTERSHOP COMMERCE MANAGEMENT REST API CONFIGURATION */
@@ -49,6 +49,7 @@ export interface Environment {
     | 'tracking'
     | 'tacton'
     | 'maps'
+    | 'sparque'
   )[];
 
   /* ADDITIONAL FEATURE CONFIGURATIONS */
